fix: tests randomly fails
