+++ admin/src/app/components/members/member-detail/member-detail.component.spec.ts
@@ -1,14 +1,8 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { initializeApp, provideFirebaseApp } from '@angular/fire/app';
-import {
-  getFirestore,
-  provideFirestore,
-  connectFirestoreEmulator,
-} from '@angular/fire/firestore';
 import { ReactiveFormsModule } from '@angular/forms';
 import { MatSnackBarModule } from '@angular/material/snack-bar';
 import { RouterTestingModule } from '@angular/router/testing';
-import { environment } from 'src/environments/environment';
+import { FirebaseTestingModule } from 'src/app/firebase-testing.module';
 import { MemberDetailComponent } from './member-detail.component';
 
 describe('MemberDetailComponent', () => {
@@ -19,14 +13,9 @@ describe('MemberDetailComponent', () => {
     await TestBed.configureTestingModule({
       imports: [
         ReactiveFormsModule,
-        provideFirebaseApp(() => initializeApp(environment.firebase)),
-        provideFirestore(() => {
-          const db = getFirestore();
-          connectFirestoreEmulator(db, 'localhost', 8080);
-          return db;
-        }),
         RouterTestingModule,
         MatSnackBarModule,
+        FirebaseTestingModule,
       ],
       declarations: [MemberDetailComponent],
     }).compileComponents();
