+++ admin/src/app/components/stores/edit-store/edit-store.component.html
@@ -42,7 +42,11 @@
       ></textarea>
     </mat-form-field>
 
-    <mat-slide-toggle formControlName="visible" class="visible">
+    <mat-slide-toggle
+      formControlName="visible"
+      class="visible"
+      ngDefaultControl
+    >
       一覧に表示する
     </mat-slide-toggle>
   </form>
