+++ admin/src/app/components/stores/edit-store/edit-store.component.spec.ts
@@ -1,5 +1,6 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-
+import { ReactiveFormsModule } from '@angular/forms';
+import { MAT_DIALOG_DATA } from '@angular/material/dialog';
 import { EditStoreComponent } from './edit-store.component';
 
 describe('EditStoreComponent', () => {
@@ -8,9 +9,10 @@ describe('EditStoreComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [ EditStoreComponent ]
-    })
-    .compileComponents();
+      declarations: [EditStoreComponent],
+      imports: [ReactiveFormsModule],
+      providers: [{ provide: MAT_DIALOG_DATA, useValue: {} }],
+    }).compileComponents();
   });
 
   beforeEach(() => {
