+++ admin/src/app/components/stores/stores.component.spec.ts
@@ -1,9 +1,7 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { initializeApp, provideFirebaseApp } from '@angular/fire/app';
-import { getFirestore, provideFirestore } from '@angular/fire/firestore';
-import { connectFirestoreEmulator } from '@firebase/firestore';
-import { environment } from 'src/environments/environment';
-
+import { MatDialogModule } from '@angular/material/dialog';
+import { MatSnackBarModule } from '@angular/material/snack-bar';
+import { FirebaseTestingModule } from 'src/app/firebase-testing.module';
 import { StoresComponent } from './stores.component';
 
 describe('StoresComponent', () => {
@@ -13,14 +11,7 @@ describe('StoresComponent', () => {
   beforeEach(async () => {
     await TestBed.configureTestingModule({
       declarations: [StoresComponent],
-      imports: [
-        provideFirebaseApp(() => initializeApp(environment.firebase)),
-        provideFirestore(() => {
-          const db = getFirestore();
-          connectFirestoreEmulator(db, 'localhost', 8080);
-          return db;
-        }),
-      ],
+      imports: [MatDialogModule, MatSnackBarModule, FirebaseTestingModule],
     }).compileComponents();
   });
