+++ admin/src/app/firebase-testing.module.ts
@@ -1,0 +1,27 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { initializeApp, provideFirebaseApp } from '@angular/fire/app';
+import {
+  getFirestore,
+  provideFirestore,
+  connectFirestoreEmulator,
+} from '@angular/fire/firestore';
+import { environment } from 'src/environments/environment';
+
+@NgModule({
+  declarations: [],
+  imports: [
+    CommonModule,
+    provideFirebaseApp(() => initializeApp(environment.firebase)),
+    provideFirestore(() => {
+      const db = getFirestore();
+      try {
+        connectFirestoreEmulator(db, 'localhost', 8080);
+      } catch (e) {
+        console.error(e);
+      }
+      return db;
+    }),
+  ],
+})
+export class FirebaseTestingModule {}
